Fix ID types of repositories

+++ src/main/java/eda/domain/DatasetRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DatasetRepository extends JpaRepository<Dataset, Integer> {
+public interface DatasetRepository extends JpaRepository<Dataset, Long> {
     Optional<Dataset> findByName(String name);
 }
